move the resource monitor back to core and reach it through fullnode.Host
